restructure project directory
