+++ internal/app/mail/datastore/mail.go
@@ -1,0 +1,1 @@
+package datastore
