+++ internal/app/mail/handler/mail.go
@@ -1,0 +1,1 @@
+package handler
