+++ internal/app/mail/interfaces/mail.go
@@ -1,0 +1,1 @@
+package interfaces
