+++ internal/app/mail/service/mail.go
@@ -1,0 +1,1 @@
+package service
